+++ modules/spring-core/src/test/kotlin/com/explyt/spring/core/completion/yaml/SpringYamlCompletionContributorTest.kt
@@ -17,22 +17,15 @@
 
 package com.explyt.spring.core.completion.yaml
 
-import com.explyt.spring.test.ExplytJavaLightTestCase
-import com.explyt.spring.test.TestLibrary
-import com.intellij.codeInsight.completion.CompletionType
-import com.intellij.codeInsight.lookup.Lookup
-import com.intellij.codeInsight.lookup.LookupManager
-import com.intellij.codeInsight.lookup.impl.LookupImpl
+import com.explyt.spring.core.completion.ExplytCompletionJavaLightTestCase
 
 
-class SpringYamlCompletionContributorTest : ExplytJavaLightTestCase() {
-
-    override val libraries: Array<TestLibrary> =
-        arrayOf(TestLibrary.springBootAutoConfigure_3_1_1, TestLibrary.springContext_6_0_7)
+class SpringYamlCompletionContributorTest : ExplytCompletionJavaLightTestCase() {
 
     override fun getTestDataPath(): String = "testdata/completion/yaml"
 
     fun testVariantsViaDot() = doTest {
+        fileName = "application.yaml"
         initSource = "sp.da.tom.nam<caret>"
         expectedLookupElements = setOf(
             "spring.datasource.tomcat.name",
@@ -43,6 +36,7 @@ class SpringYamlCompletionContributorTest : ExplytJavaLightTestCase() {
     }
 
     fun testInsertToExistKey() = doTest {
+        fileName = "application.yaml"
         initSource = """
 sprindatatomna<caret>
 
@@ -77,6 +71,7 @@ spring:
     }
 
     fun testInsertToExistKeyFromOneLevelParent() = doTest {
+        fileName = "application.yaml"
         initSource = """
 spring:
   datasource:
@@ -112,6 +107,7 @@ spring:
 
 
     fun testSimpleInsertToKey() = doTest {
+        fileName = "application.yaml"
         initSource = """
 spring:
   datasource:
@@ -141,6 +137,7 @@ spring:
     }
 
     fun testInsertToExistKeyToDown() = doTest {
+        fileName = "application.yaml"
         initSource = """
 spring:
   datasource:
@@ -170,34 +167,5 @@ spring:
       """.trimIndent()
     }
 
-    private fun doTest(init: TestModel.() -> Unit) {
-        val model = TestModel()
-        model.init()
-
-        myFixture.configureByText("application.yaml", model.initSource)
-
-        val lookupElements = myFixture.complete(CompletionType.BASIC)
-        assertNotNull(lookupElements)
-
-        val lookupElementStrings = lookupElements.map { it.lookupString }
-
-        assertEquals(
-            model.expectedLookupElements, lookupElementStrings.toSet()
-        )
-
-        getActiveLookup().currentItem = lookupElements[0]
-        getActiveLookup().finishLookup(Lookup.NORMAL_SELECT_CHAR)
-        myFixture.checkResult(model.sourceAfterComplete)
-    }
-
-    private fun getActiveLookup(): LookupImpl {
-        return LookupManager.getActiveLookup(myFixture.editor) as LookupImpl
-    }
-
-    private class TestModel {
-        lateinit var initSource: String
-        lateinit var sourceAfterComplete: String
-        lateinit var expectedLookupElements: Set<String>
-    }
 }