+++ modules/spring-core/src/test/kotlin/com/explyt/spring/core/providers/java/SpringLineMarkerProviderTest.kt
@@ -40,7 +40,7 @@ class SpringLineMarkerProviderTest : ExplytJavaLightTestCase() {
 
         myFixture.doHighlighting()
 
-        val lineMarkers = DaemonCodeAnalyzerImpl.getLineMarkers(myFixture.editor.document, project);
+        val lineMarkers = DaemonCodeAnalyzerImpl.getLineMarkers(myFixture.editor.document, project)
 
         assertEquals(3, lineMarkers.size)